Don't delete the concrete field unless there are no more types using it

+++ rea/behaviour/identifier/model.py
@@ -117,8 +117,11 @@ class IdentificationField(models.Model):
         for vf in self:
             field = vf.field
             super(IdentificationField, self).unlink()
-            # TODO prevent deleting existing data
-            field.unlink()
+            # delete the concrete field if there are no remaining identfields
+            if not self.search([('field_name', '=', field.name),
+                                ('model', '=', field.model_id.id)]):
+                # TODO prevent deleting existing data
+                field.unlink()
 
 
 class Identification(models.Model):
